restricoes-permissoes

+++ index.php
@@ -2,24 +2,26 @@
 include 'inc/header.php';
 include 'classes/contatos.php';
 include 'classes/funcoes.php';
+require_once 'classes/usuario.php';
 
 $contato = new Contato();
 $fn = new Funcoes();
+$usuario = new Usuario();
 
 session_start();
-if (!($_SESSION)) {
+if (!isset($_SESSION['logado'])) {
     header("Location: login.php");
+    exit;
 }
+
+$usuario->setUsuario($_SESSION['logado']);
 ?>
 <main>
-    <!-- <?php if ($_SESSION): ?>
+    <?php if ($usuario->temPermissao("add")): ?>
         <div class="button">
-            <button><a href="#">Sair</a></button>
+            <button><a href="adicionarContato.php">Adicionar um contato</a></button>
         </div>
-    <?php endif; ?> -->
-    <div class="button">
-        <button><a href="adicionarContato.php">Adicionar um contato</a></button>
-    </div>
+    <?php endif; ?>
     <table class="tabela-contatos">
         <thead>
             <tr>
@@ -33,7 +35,9 @@ if (!($_SESSION)) {
                 <th>NASCIMENTO</th>
                 <th>FOTO</th>
                 <th>ATIVO</th>
-                <th>AÇÕES</th>
+                <?php if ($usuario->temPermissao("edit") || $usuario->temPermissao("del")): ?>
+                    <th>AÇÕES</th>
+                <?php endif; ?>
             </tr>
         </thead>
         <?php
@@ -52,17 +56,30 @@ if (!($_SESSION)) {
                     <td><?php echo $fn->dtNasc($item['dtNasc'], 2); ?></td>
                     <td><?php echo $item['foto']; ?></td>
                     <td><?php echo $item['ativo']; ?></td>
-                    <td class="acoes">
-                        <a href="editarContato.php?id=<?php echo $item['id'] ?>">EDITAR </a>
-                        |
-                        <a href="excluirContato.php?id=<?php echo $item['id'] ?>" onclick="return confirm('Você tem certeza que quer excluir esse contato?')"> EXCLUIR</a>
-                    </td>
+                    <?php if ($usuario->temPermissao("edit") || $usuario->temPermissao("del")): ?>
+                        <td class="acoes">
+                            <?php if ($usuario->temPermissao("edit")): ?>
+                                <a href="editarContato.php?id=<?php echo $item['id'] ?>">EDITAR </a>
+                            <?php endif; ?>
+                            <?php if ($usuario->temPermissao("edit") && $usuario->temPermissao("del")): ?>
+                                |
+                            <?php endif?>
+                            <?php if ($usuario->temPermissao("del")): ?>
+                                <a href="excluirContato.php?id=<?php echo $item['id'] ?>" onclick="return confirm('Você tem certeza que quer excluir esse contato?')"> EXCLUIR</a>
+                            <?php endif; ?>
+                        </td>
+                    <?php endif; ?>
                 </tr>
             </tbody>
         <?php
         endforeach;
         ?>
     </table>
+    <?php if ($usuario->temPermissao("super")): ?>
+    <div class="button">
+        <button><a href="infoUser.php">Usuario</a></button>
+    </div>
+    <?php endif; ?>
 </main>
 
 <?php include 'inc/footer.php'; ?>
